remove custom logger and merge with old logging code
