+++ util/log.go
@@ -7,17 +7,7 @@ import (
 	log "github.com/sirupsen/logrus"
 	"gopkg.in/natefinch/lumberjack.v2"
 
-	"github.com/netbirdio/netbird/management/server/context"
-)
-
-type LogSource string
-
-const (
-	LogSourceKey = "logSource"
-
-	HTTPSource   LogSource = "HTTP"
-	GRPCSource   LogSource = "GRPC"
-	SystemSource LogSource = "SYSTEM"
+	"github.com/netbirdio/netbird/formatter"
 )
 
 // InitLog parses and sets log-level input
@@ -40,85 +30,7 @@ func InitLog(logLevel string, logPath string) error {
 		log.SetOutput(io.Writer(lumberjackLogger))
 	}
 
-	log.SetFormatter(NewCustomFormatter())
+	formatter.SetTextFormatter(log.StandardLogger())
 	log.SetLevel(level)
 	return nil
-}
-
-// CustomFormatter formats the log message as required
-type CustomFormatter struct {
-	log.TextFormatter
-}
-
-// NewCustomFormatter creates a new CustomFormatter
-func NewCustomFormatter() *CustomFormatter {
-	formatter := CustomFormatter{}
-	formatter.TextFormatter.DisableSorting = true
-	return &formatter
-}
-func (f *CustomFormatter) Format(entry *log.Entry) ([]byte, error) {
-	if entry.Context == nil {
-		return f.TextFormatter.Format(entry)
-	}
-	source, ok := entry.Context.Value(LogSourceKey).(LogSource)
-	if !ok {
-		return f.TextFormatter.Format(entry)
-	}
-
-	entry.Data["source"] = source
-	switch source {
-	case HTTPSource:
-		return f.formatHTTPLog(entry)
-	case GRPCSource:
-		return f.formatGRPCLog(entry)
-	case SystemSource:
-		return f.formatSystemLog(entry)
-	default:
-		return f.TextFormatter.Format(entry)
-	}
-}
-
-func (f *CustomFormatter) formatHTTPLog(entry *log.Entry) ([]byte, error) {
-	if ctxReqID, ok := entry.Context.Value(context.RequestIDKey).(string); ok {
-		entry.Data[context.RequestIDKey] = ctxReqID
-	}
-	if ctxAccountID, ok := entry.Context.Value(context.AccountIDKey).(string); ok {
-		entry.Data[context.AccountIDKey] = ctxAccountID
-	}
-	if ctxInitiatorID, ok := entry.Context.Value(context.UserIDKey).(string); ok {
-		entry.Data[context.UserIDKey] = ctxInitiatorID
-	}
-
-	return f.TextFormatter.Format(entry)
-}
-
-func (f *CustomFormatter) formatGRPCLog(entry *log.Entry) ([]byte, error) {
-	if ctxReqID, ok := entry.Context.Value(context.RequestIDKey).(string); ok {
-		entry.Data[context.RequestIDKey] = ctxReqID
-	}
-	if ctxAccountID, ok := entry.Context.Value(context.AccountIDKey).(string); ok {
-		entry.Data[context.AccountIDKey] = ctxAccountID
-	}
-	if ctxDeviceID, ok := entry.Context.Value(context.PeerIDKey).(string); ok {
-		entry.Data[context.PeerIDKey] = ctxDeviceID
-	}
-
-	return f.TextFormatter.Format(entry)
-}
-
-func (f *CustomFormatter) formatSystemLog(entry *log.Entry) ([]byte, error) {
-	if ctxReqID, ok := entry.Context.Value(context.RequestIDKey).(string); ok {
-		entry.Data[context.RequestIDKey] = ctxReqID
-	}
-	if ctxInitiatorID, ok := entry.Context.Value(context.UserIDKey).(string); ok {
-		entry.Data[context.UserIDKey] = ctxInitiatorID
-	}
-	if ctxAccountID, ok := entry.Context.Value(context.AccountIDKey).(string); ok {
-		entry.Data[context.AccountIDKey] = ctxAccountID
-	}
-	if ctxDeviceID, ok := entry.Context.Value(context.PeerIDKey).(string); ok {
-		entry.Data[context.PeerIDKey] = ctxDeviceID
-	}
-
-	return f.TextFormatter.Format(entry)
 }
